Wrap qualified identifier in a sub-tree
By wrapping in a subtree we can use qualified_identifier.text to extract all the input that was matched by that rule, otherwise
'text' only matches the first node in a list.  Took me ages to find that :(

Send comments to hidden channel, hiopefully we can then retrieve them later.